Move MessengerExtensions to ipc module
