+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/serviceconnection/CustomDns.kt
@@ -3,9 +3,9 @@ package net.mullvad.mullvadvpn.ui.serviceconnection
 import android.os.Messenger
 import java.net.InetAddress
 import net.mullvad.mullvadvpn.lib.ipc.Request
+import net.mullvad.mullvadvpn.lib.ipc.extensions.trySendRequest
 import net.mullvad.mullvadvpn.model.DnsOptions
 import net.mullvad.mullvadvpn.model.DnsState
-import net.mullvad.mullvadvpn.util.trySendRequest
 import net.mullvad.talpid.util.EventNotifier
 
 class CustomDns(private val connection: Messenger, private val settingsListener: SettingsListener) {
