+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/serviceconnection/ServiceConnectionDeviceDataSource.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.callbackFlow
 import net.mullvad.mullvadvpn.lib.ipc.Event
 import net.mullvad.mullvadvpn.lib.ipc.EventDispatcher
 import net.mullvad.mullvadvpn.lib.ipc.Request
-import net.mullvad.mullvadvpn.util.trySendRequest
+import net.mullvad.mullvadvpn.lib.ipc.extensions.trySendRequest
 
 class ServiceConnectionDeviceDataSource(
     private val connection: Messenger,
